Real-time point visualization test
Added preliminary support for both real-time point visualization for the included eye-tracking data, as well as mouse-look for the scene camera so the user can look around the scene at the projected data. Static data (in red) is still available and can be shown either alongside real-time data, or indepenantly.